refactor: 공통 패키지를 생성하라
- PurgomalumClient 이동
