+++ src/main/java/kitchenpos/common/application/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.common.application;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
