+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,6 +1,6 @@
 package kitchenpos.product.application;
 
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.common.application.PurgomalumClient;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
