+++ src/test/java/kitchenpos/FakePurgomalumClient.java
@@ -1,6 +1,7 @@
 package kitchenpos;
 
-import kitchenpos.infra.PurgomalumClient;
+
+import kitchenpos.common.application.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
